refactor: create floating card component
